feat(ui): add ConfigureSSO navbar and skeleton, refine breadcrumb sizing
Splits the navbar chrome into ConfigureSSONavbar so the layout
composition is more readable. Adds a ConfigureSSOSkeleton that
mirrors the live shell — breadcrumb placeholders, centered content
spinner, disabled footer buttons — used by the initial-load gate.

Drops the orphaned ConfigureSSOCard (replaced by inline
ProfileCard.Root + the new navbar component). Adjusts breadcrumb
bullet sizing and the separator caret to match the latest visual
direction.
